Ugly hack for fixing the multiline support in LatexMath
Newlines should work with '\\' at the end of the line but this is not
the case. This tiny hack although is not clean, it allows me to move fast.
As for now the only place where we're using this kind of math syntax is
in just one file (implemented-voting-schemes/dummy.adoc), then I can
live with it.